User book relation

+++ src/controllers/books/books.controller.ts
@@ -2,12 +2,14 @@ import {
 	Body,
 	Controller,
 	Delete,
+	ForbiddenException,
 	Get,
 	InternalServerErrorException,
 	NotFoundException,
 	Param,
 	Post,
 	Put,
+	Request,
 } from '@nestjs/common';
 import {BooksService} from 'src/services/books/books.service';
 import {
@@ -15,9 +17,13 @@ import {
 	ApiOkResponse,
 	ApiCreatedResponse,
 	ApiNotFoundResponse,
+	ApiForbiddenResponse,
 } from '@nestjs/swagger';
 import {BookDto} from 'src/dtos/books/Book.dto';
 import {CreateBookDto} from 'src/dtos/books/CreateBook.dto';
+import {UserRole} from 'src/typeorm/entities/User';
+import {JwtSignPayload} from 'src/dtos/auth/LoginPayload.dto';
+import {Book} from 'src/typeorm/entities';
 
 @ApiTags('books')
 @Controller('books')
@@ -30,14 +36,19 @@ export class BooksController {
 		type: BookDto,
 		description: 'Requested books',
 	})
-	get(): Promise<BookDto[]> {
-		return this.bookService.get();
+	get(@Request() req): Promise<BookDto[]> {
+		const {user} = req;
+		const authorId = user.role === UserRole.ADMIN ? user.id : undefined;
+		return this.bookService.get(authorId);
 	}
 
 	@Get(':id')
 	@ApiOkResponse({type: BookDto, description: 'Requested book'})
-	getById(@Param('id') id: string): Promise<BookDto> {
-		return this.bookService.getById(id);
+	@ApiForbiddenResponse({description: 'User does not have permission for this action'})
+	async getById(@Param('id') id: string, @Request() req): Promise<BookDto> {
+		const book = await this.bookService.getById(id);
+		await this.validateBookOwnership(req.user, book.authorId);
+		return book;
 	}
 
 	@Post()
@@ -45,18 +56,26 @@ export class BooksController {
 		type: BookDto,
 		description: 'Book successfully created',
 	})
-	create(@Body() book: CreateBookDto): Promise<BookDto> {
+	@ApiForbiddenResponse({description: 'User does not have permission for this action'})
+	async create(
+		@Body() book: CreateBookDto,
+		@Request() req
+	): Promise<BookDto> {
+		await this.validateBookOwnership(req.user, book.authorId);
+
 		return this.bookService.create(book);
 	}
 
 	@Put(':id')
 	@ApiOkResponse({description: 'Book sucessfully updated'})
 	@ApiNotFoundResponse({description: 'Requested book not found'})
+	@ApiForbiddenResponse({description: 'User does not have permission for this action'})
 	async update(
 		@Param('id') id: string,
-		@Body() book: CreateBookDto
+		@Body() book: CreateBookDto,
+		@Request() req
 	): Promise<void> {
-		await this.checkBookExistance(id);
+		await this.validateExistanceAndOwnership(id, req.user);
 
 		const result = await this.bookService.update(id, book);
 		if (!result) throw new InternalServerErrorException();
@@ -67,8 +86,9 @@ export class BooksController {
 	@Delete(':id')
 	@ApiOkResponse({description: 'Book sucessfully deleted'})
 	@ApiNotFoundResponse({description: 'Requested book not found'})
-	async delete(@Param('id') id: string): Promise<void> {
-		await this.checkBookExistance(id);
+	@ApiForbiddenResponse({description: 'User does not have permission for this action'})
+	async delete(@Param('id') id: string, @Request() req): Promise<void> {
+		await this.validateExistanceAndOwnership(id, req.user);
 
 		const result = await this.bookService.delete(id);
 		if (!result) throw new InternalServerErrorException();
@@ -76,9 +96,27 @@ export class BooksController {
 		return;
 	}
 
-	private async checkBookExistance(id: string): Promise<void> {
+	private async checkBookExistance(id: string): Promise<Book> {
 		const existingBook = await this.bookService.getById(id);
 		if (!existingBook)
 			throw new NotFoundException(`Book with identifier ${id} not found`);
+
+		return existingBook;
+	}
+
+	private async validateBookOwnership(
+		user: JwtSignPayload,
+		authorId: string
+	): Promise<void> {
+		if (user.role !== UserRole.ADMIN || authorId !== user.id)
+			throw new ForbiddenException();
+	}
+
+	private async validateExistanceAndOwnership(
+		id: string,
+		user: JwtSignPayload
+	): Promise<void> {
+		const book = await this.checkBookExistance(id);
+		this.validateBookOwnership(user, book.authorId);
 	}
 }
